Simple update for Level 2 (50% success)
A simple update to the original code that inserts a check to see if a floor has an elevator. If it hasn't, the leading clone is instructed to build one. This passes the fist 5 tests but fails all the subsequent ('best path') tests. These look difficult. The drives could me modelled as graphs, with nodes and edges etc. The natural direction of the clones would need to be taken into account. For example, a single node to the left entry elevator to a floor would need to have a movement cost of 3 rather than 1 (since a 'turn'[move, block, move] is needed to reach it). In this way, a best path could be calculated. But later puzzles require the building of elevators to construct a best path. I don't know how to do this. Could this be done iteratively? Eg: try all the possible placements of x elevators and store the best path generated? It feels like the whole code needs quite a big re-write.